+++ CHANGELOG.md
@@ -10,6 +10,12 @@ All notable changes to this project will be documented in this file. See
 - harden queue semantics and migrate to oxc tooling ([ce5856c](https://github.com/calebboyd/aceink/commit/ce5856c2df1b6b9fc73b13dd7128948c398839fc))
 - update workflows ([0ecd55a](https://github.com/calebboyd/aceink/commit/0ecd55a7e836ce2b4172cbab2ddcfb51878ecbd4))
 
+## [Unreleased]
+
+### Breaking Changes
+
+- fold queue `bound` into the `QueueOptions` object for both `new Queue()` and `q()`
+
 # [2.0.0](https://github.com/calebboyd/async/compare/v1.3.0...v2.0.0) (2026-03-16)
 
 ### Breaking Changes